Langchain/tools/requirements.txt
langchain
pydantic
langchain_core
python-dotenv
langchain-google-genai
google-genai